Remove outer Context provider
It turns out this is simpler than I had figured initially.

LocationCache will (a) load in a location the first time,
and then (b) cache it and always return from the cache.

This means we need to call it (a) once on the server,
and then (b) once on the client.

Anything passed through `getInitialProps` is available (a)
on the server, and then (b) once on the client.
(ie, props.context is available on the first page render,
but then we don't have it anymore as we move between pages).

We can pass a subdomain string through from the server and
it will be available on the server and the client. Once
LocationCache is called once on each environment, it saves
the location and we don't need to reset it again.

So we never need to inspect the subdomain path on the client,
nor do we need to attach a global variable to window or
anything like that.