Allow location objects to be passed in
One of our cypress tests is breaking because the require path
is relative to the app directories.

By allowing us to pass in an object instead of a string, we can
manually pass in a location object in the test instead of
trying to do path introspection or something horrible like that.

Added a comment in the cypress test explaining what the
getGlobalLocation call is doing.

Also renamed a few test files because of small mistakes in their
filenames.