Lingui upgrade (#402)
* initial package upgrade

* message compile + test updates

* Title HOC

* fix NoJS test

* Rename lingui key in package.json

* Small CSS fixes now that spans are gone

- more spacing under radio buttons
On the register page, our spans had some bottom-margins applied
that no longer exist. Added more bottom padding to the label so
that they're less squished.

- use margin instead of padding for chevron on start page
The arrow was an image with a max-width and border box. Adding
padding to it kept the width fixed but pushed the image into a
smaller space, so the whole arrow shrunk as a result.
Using margin instead of padding fixes this.