Create Language component
React context contains pretty much all of the session data we have
throughout the app, but it's not always easy to get.

If all we need is the language string, wrapping another component
in Language, and passing the language string as a render prop is
quite a bit easier than doing a large object destructure every time.