Stop infinite loop for startup errors
In our middleware, we will throw an error if there is no location
string or if there is no receiving email address shown.

On the landing page, if we hit an error, it would redirect to the error page.
Unfortunately, the error page would also encounter the same error.

This meant that we would end up with an infinite loop if we
ran into any of these.

This commit generalises some of the code, and helps mitigate
infinite error loops.