Remove contact link on generic not found page (#410)
If redirecting from the server (i.e. no location found) we need to have a generic not found page.  This is needed to avoid people contacting an incorrect office.

Checks are based on a specific page name /not-found

Other 404 pages /some-page will be treated as a location not found page and still contain the contact link