Move error out of LocationCache
With the Error being thrown in LocationCache, any call in the
application could potentially trigger it depending on the order
of code execution.

It's possible to be much more prescriptive about this.

- We want to set the location once upfront.
- If no location string exists and
  if no location has been saved already
  then we want to thrown an error

Removed the error from LocationCache and put it into a
new function called `setGlobalLocation`, which is called
at the very beginning when the app starts up.

If en error is thrown there, then we go to the 500 page.
Once we're on the 500 page, we won't throw the LocationCache
error anymore so hopefully we no longer have to worry about
infinite loops that send 70k error messages to sentry.
What we want is to set the location once, and then be able to not
throw that error again once we have redirected to the error page.