Add function to remove empty arrays from form values
Super weird behaviour for the checkbox on the register page.

Before the checkbox is clicked, values:
`{}`

Once the checkbox is clicked, values:
`{ familyCheck: ['familyCheck'] }`

Once the checkbox is unclicked, values:
`{ familyCheck: [] }`

So we can't use `required_if`. This will work if we _never_ check
the family option, but it won't work if we check and then uncheck it.
Once it sees `{ familyCheck: [] }`, it thinks that family check
has a real value, even though it doesn't.

But then, because of the way validatorjs works, we can't use
`required_with` to match against arrays.
ie, the check it will perform is
['familyCheck'] === ['familyCheck'] // returns false

So, rather than doing a custom validation rule -- not too sure how
this would work, especially since we need to check against the
value of another field -- we can delete the key entirely from the
values object if the array is empty.