Error message for familyCheck
Added an error message for the family check.
Did a slightly weird thing of giving the error message the id
that will show up in the error message box at the top of the page.

It's super weird to have to check the family member box to clear
the data and then uncheck it again, but it's late.