Don't make sure all the fields are in query keys
Since we sometimes have optional keys, we don't need to make sure
that all keys exist in the query parameters.
What we do want to do though is make sure that all keys that are
in the query are also in the page fields. (ie, they can be a subset,
but we don't accept anything we don't recognize).