WIP - Family Option (#404)
* New Features: Registration NO VALIDATION YET

* Increased checkbox border by 1px on Reg Page

* stopping for the day

* Validation, values populating review page

* Entry Point for Paul

* Text Update + Styling Update

* Add function to remove empty arrays from form values

Super weird behaviour for the checkbox on the register page.

Before the checkbox is clicked, values:
`{}`

Once the checkbox is clicked, values:
`{ familyCheck: ['familyCheck'] }`

Once the checkbox is unclicked, values:
`{ familyCheck: [] }`

So we can't use `required_if`. This will work if we _never_ check
the family option, but it won't work if we check and then uncheck it.
Once it sees `{ familyCheck: [] }`, it thinks that family check
has a real value, even though it doesn't.

But then, because of the way validatorjs works, we can't use
`required_with` to match against arrays.
ie, the check it will perform is
['familyCheck'] === ['familyCheck'] // returns false

So, rather than doing a custom validation rule -- not too sure how
this would work, especially since we need to check against the
value of another field -- we can delete the key entirely from the
values object if the array is empty.

* Error message for familyCheck

Added an error message for the family check.
Did a slightly weird thing of giving the error message the id
that will show up in the error message box at the top of the page.

It's super weird to have to check the family member box to clear
the data and then uncheck it again, but it's late.

* Conditionally add family members to summary page

* Remove field nesting

* only disable textArea if JS enabled

* reload checkbox

* fix disabled state for textArea using didMount

* ensure form options are passed

* test update ensure error shows

* Add family NoJS test

* Updated Staff + Client Email

* Removed duplicate label + fixed textArea css

* Email Update

* Reformatted email

* Validation improved for family options

The behaviour we want is:

- familyCheck is optional
- familyOption is optional if familyCheck is not been checked
- familyOption is required if familyCheck has been checked

Should work the same with or without JS.

* Don't make sure all the fields are in query keys

Since we sometimes have optional keys, we don't need to make sure
that all keys exist in the query parameters.
What we do want to do though is make sure that all keys that are
in the query are also in the page fields. (ie, they can be a subset,
but we don't accept anything we don't recognize).

* Improve markup for the family checkbox and textarea

In brief:

- removed validation message and validaion label for familyCheck
  since that will never throw a validation error
- moved the validation message for the familyOption into the label
  for the textarea

* Clear familyOption field if familyCheck is absent

We only care about the familyOption input if familyCheck exists.
If familyCheck doesn't exist, we can just clear whatever they
had in there.

Easy as.

* Clean up CSS for checkbox and textarea

A few things in here:
- generalised the disabled rules so that any textareas or text
  inputs can use them
- changed theme.colour.grey since nobody is using it
- made CSS more explicit in a few cases (is, using border-width or
  background-color instead of setting all properties)
- made checkbox label inline-block so that people don't click or
  unclick it accidentally by clicking whitespace
- reorganized CSS rules so that targetted elements are following
  the same page layout

* Lint + reorder some parameters

Fixed linting problems, and reordered parameters in a couple places
to keep them in the order we have them on the register page.

* Fixed noJS-family tests

Also updated the no-js-jest config so that yarn test still workds.

* Remove comma from list of names in applicant email

* Modify NoJS tests

* Add missing french translations

Also remove error message for familyCheck that will never be called.