Add clickAndWait function
We were seeing a lot of intermittent failures when running
`yarn test:full` with the puppeteer tests giving us an
`Execution context was destroyed` error.

Seems like that error message is trying to tell us that the
page is being navigated away from in a way that puppeteer
doesn't like.
- https://github.com/GoogleChrome/puppeteer/issues/1591#issuecomment-355245777

Someone suggested using page.waitFor because that solved
their issue. Seems to work for our case.
- https://github.com/GoogleChrome/puppeteer/issues/1591#issuecomment-363384982

Created a new function `clickAndWait` that will call
page.click and then page.waitFor.
Running this a bunch of times locally and our tests
seems to pass reliably now.