Only run tests for `*.test.js` files
Splitting out some of our puppeteer test setup and I was getting
an error that it was trying to run tests in `puppeteer-utils.js`
file but not finding any tests to run.

Since all of our tests files are called `*.test.js`, it seemed
fine to make our search a bit more strict.