Make sure familyCheck is not an empty string
In NoJS mode, not checking the checkbox returns a query string
of `familyCheck=`, which translates back into an empty string.

Since we were only checking if familyCheck was a string, this meant
we would always check the checkbox after validating.

Now, we make sure it's not an empty string before prefilling it.