Require familyCheck if familyOption exists by default
In JS mode, we can disable the textarea which prevents people
from interacting with it as well as visually de-emphasizes it.
So if someone enters their family members but then unchecks the
box, it's a reasonable assumption that they've done that
consciously.

In noJS mode, we don't have a way to disable the box, so it would
be much easier for someone to fill in some family members without
clicking the checkbox.

Therefore, we're going to have a validation rule that runs on the
noJS version requiring the family checkbox is checked, but ignore it
on the JS version.