Add unavailability flow checkbox
Rather than a link, we are going to have a checkbox that users
click if they are not available. Once clicked, the calendar will
be visually disabled and then the "submit" button at the bottom
of the page will link to the unavailability explanation flow
rather than the review page.

We are also only going to validate selectedDays under the `submit`
route if no explanationPage value exists.
If it exists at all, we assume that the user is submitting an
explanation string rather than a list of dates.