Upload sourcemaps from circleci after container deployment
Since moving to containerized deploys instead of shipit deploys,
we lost our sourcemap uploads to sentry.

This PR adds a step to circleci which will upload the sourcemaps
after the container deploys.

Added a new environment variable called SENTRY_AUTH_TOKEN to
circleci so that the uploads work.
Currently using _my_ auth token, so we can change that if needed.