Announce dates as unavailable
When using a screen reader dates were being read and than announce as dimmed when a date is disabled.  The user would need to listen to the entire date in order to find out if it was available.

This update announces disabled dates as "unavailable" prior to reading the date.