Submission form may get empty string for `selectedDates`
If dates haven't been set, the submission form gets an empty string
for selectedDates -- this was giving us a warning in the console.

In other places in the app, we're setting selectedDates to be an
empty array by default, so maybe it would be better to do that.
However, since 'truthy' checks in JS would consider an empty
string to be 'false' but an empty array to be 'true', I figured I
would leave it as an empty string and instead expand what PropTypes
are allowed.