Set global 'extra' context for raven
For our manually triggered Raven errors
(ie, `Raven.captureException` or `Raven.captureMessage`), Raven
isn't capturing all the contextual data automatically, the
way it does when it captures a `throw new Error` event.

What this means is that we have a lot of errors showing up in
sentry because our tests trigger them, but we're not able to
at-a-glance know when they're coming from our production
application.

By setting a global context.extra, all exceptions and messages
captured will have the following useful context:
- the url host name
- the subdomain
- the url path
- the release string
- the RAZZLE_STAGE env var