Add RAZZLE_FLAGS at build time
Any environment variable that needs to be accessed in the client
needs to be passed into the container at build time.

To add a new flag variable to the app, we have to update the
RAZZLE_FLAGS environment variables in:

- CircleCI (for the client-side variable to exist)
- the AWS Fargate task definition (for the server-side variable to exist)

This is cutting-edge tech, so we will obviously document this.