Redirect can be a getter again
When the `next page` link was dynamic, we needed some logic that
would swap the redirect link out for the other page based on
whether or not there was an explanation page value.

Now that we are making the lunks un-dynamic again, we can
revert redirect back to being a getter like it was before.

C'est plus simple comme ca.

More on getters:
- https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/get