Add explanation page field to CalendarFields
Updated `setStore` to reset `explanationPage` from the calendar page.

Initially ran into a problem where it refused to reset the value.

Here was the problem:
- when we call `setStore`, we have to pass in an object (ie, `{explanationPage: ''}`)
- this object is run through `validateCookie` before it is saved
  - this is how we make sure that all saved values are validated first
- validateCookie uses `fields` and `validate` from the current page
  and runs them against the object that was passed in
- Since `explanationPage` was not part of the Calendar's fields,
  it was originally refusing to validate it and then it wouldn't
  reset the value.
- By adding `explanationPage` in as a calendar field with the
  'accept_anything' validation rule, now the validation _will_ pass
  and the value will be reset.

Note: when validating fields manually, fields that are not included
in the validation call will be reset to empty strings.
This means that when we call validateCookie on `{explanationPage: ''}`,
it returns an object with all of the calendar fields:
`{explanationPage: '', selectedDays: ''}`
This means that we end up with an empty selectedDays key in the cookie
under `explanation`. It's not needed but it also doesn't have any
untoward effect on the rest of our code so we're safe.