Generate unique id attributes for cancel button svg titles
Flagged by axe: ids should be unique on a page and we were
using duplicates for the <title> element inside of the
"Remove day" button svg.

Looked into writing a pa11y test for this, but it was not
going to be straightforward, so just going to add the fix
by itself.