Add saveAfter static function to pages
Since we have a situation now where we have to set values for another
page, I added a new function that will save something else to the
store if the original validation works.

In our specific case:
- whenever we save anything from the calendar page
- we want to clear the explanationPage value

Doesn't matter if it exists already or not -- we just want to set it
to an empty string all of the time.

Note:
setting values this way breaks our previous conventions
and eschews the validation we have in place in all other cases, but
- it's hardcoded, so there's no user-input problem
- and we have to get this shipped pretty soon.

There's probably a nicer way to do this but this is cheap and
effective.