Pass current context to saveAfter function
It turns out that on the calendar page in JS mode, we don't want
to erase the explanation page values if we submit with the
'availability' checkbox checked.

By passing the calendar values back into the saveAfter function, we
can check if the availability checkbox was checked, and then, if it
was return an empty object.

In `WithProvider`, we won't save anything if we have empty objects.