Ignore generated translations files
All of the translations files in `locale/_build` are generated by
lingui and not something we actually need to commit to the repo.

They make the diffs really noisy and they aren't files we ever have
to save or modify.

We need to keep the `en` stuff at present because it's imported
in src/locale/linguiUtils.js, but with a better refactor we could
probably get rid of it.

This will make future development easier and it would have made
past development easier as well.
Should have done this months ago.