Convert passed-in dates to ISO strings before building noJS calendar
Very minor bug here, arguably not a bug at all.

If you selected days on the no-JS calendar, and then turned on JS
and reloaded the page, you would see your days on the JS calender.

But doing the opposite, (loading no-JS cal after selecting them on
the JS cal) wasn't working. This is because the date format that
was being passed through was a timestamp and then the
`dates.includes(val)` check wouldn't ever match.

Converting the array passed-in dates works as expected.