Remove Heroku deployment config
Now that the app is being handed back to IRCC, we don't anticipate
they will need or use Heroku for anything.

Leaving it in wouldn't hurt anything actively (it would still work),
but removing it means we have less logic to manage in total.

Outcome: our service is simpler to run and understand.

The setup can still be referenced by looking at what was removed
in this commit.