Update puppeteer test config
Was getting some failing tests for the puppeteer no-js-family tests,
although just running the app, I was seeing the correct behaviour.
Also running the puppeteer tests locally was fine, so this was a
"see if it works on circleci" process of debugging.

Made two changes:
- updated the "wait and click" duration time to 1 second
- changed the family field noJS error tests to just use URLs instead of
  actually mocking out the behaviour

The family field error tests were the ones that were causing the issue,
but since it's untestable locally, I figured I would simplify them
so that they're more straightforward.