Don't show "Phone -" on 500 page if no location
Our root domain has no location set, so we don't have contact
information to show users.

Right now, it will say "Phone -" and then not have a phone number.
Instead, we should omit the Contact output entirely if there is no
contact info.

Notes:
1. Added some mocks to the previously unmocked "getEmail" function
I was relying on.

2. Had to do some funny renaming stuff so that the default export
didn't line up with the name of one of the named exports, but
functionally there is no difference.