Fix horizontal padding throughout the app
After upgrading to emotion 10, our horizontal padding media queries
broke for the main content but not for other parts of the app
(for example, the header and footer were fine).

So I've collected all the horizontal padding rules together and applied
them uniformly.

Now they will all work or they will all break as a team.