Remove "ensure receiving email" logic on every request
Since the email that we want to send the request to will vary location
by location, it is possible we might get into a position where
we try to send an email but don't have one configured.

However, the logic needed to do this check has become increasingly
cumbersome. For example,
- we want to check if we are in prod or staging
- if staging, we want to check for the 'test' email env var
- if prod, we want to check for the location being set
- and now we would also have to check if we are at the root
domain or not

We could either keep piling logic into our `getReceivingEmail` method,
OR, accepting the risk, we could remove the check to make sure we
have an email at the start of every request.

This means that a missing email would only be discovered
when someone is about to submit rather than on app startup, which would
be a worse experience than now.

However:
- when this happens we throw an error and it ends up in Sentry
  - we have never seen this error show up in Sentry except on developer laptops
- we have since added tests whereby all location files need to have an email
defined

So I am comfortable removing this check.