Skip puppeteer noJS tests on CI
Something about our setup keeps failing these specific tests
on CircleCI, and I have trial and errored them, but at this
point I am not sure what the problem is and it is resulting
in a flakey deployment pipeline.

There have been a couple attempts to ameliorate this in the
past but these tests continue to fail intermittently.

Biting the bullet and removing them from CircleCI.

They can still be run locally with `yarn test:full`